move and exclude expected results files
