+++ .codecov.yml
@@ -25,6 +25,7 @@ github_checks:
 ignore:
   - ".github/"
   - ".binder/"
+  - "aeon/tests/expected_results/"
   - "build_tools/"
   - "docs/"
   - "examples/"
