+++ aeon/classification/tests/test_all_classifiers.py
@@ -7,12 +7,13 @@ import inspect
 import numpy as np
 from sklearn.utils._testing import set_random_state
 
-from aeon.classification.tests.test_expected_outputs import (
+from aeon.datasets import load_basic_motions, load_unit_test
+from aeon.testing import (
+    BaseFixtureGenerator,
+    QuickTester,
     basic_motions_proba,
     unit_test_proba,
 )
-from aeon.datasets import load_basic_motions, load_unit_test
-from aeon.tests.test_all_estimators import BaseFixtureGenerator, QuickTester
 from aeon.utils._testing.estimator_checks import _assert_array_almost_equal
 from aeon.utils._testing.scenarios_classification import ClassifierFitPredict
 from aeon.utils.validation.collection import get_n_cases
