+++ aeon/datasets/tests/test_load_forecasting.py
@@ -7,7 +7,7 @@ from pandas.testing import assert_frame_equal
 
 import aeon
 from aeon.datasets import load_forecasting, load_from_tsf_file, load_uschange
-from aeon.tests.test_config import PR_TESTING
+from aeon.testing import PR_TESTING
 
 _CHECKS = {
     "uschange": {
