+++ aeon/testing/__init__.py
@@ -1,0 +1,1 @@
+"""Package wide tests and testing utilities for aeon."""
