+++ aeon/testing/expected_results/__init__.py
@@ -1,0 +1,1 @@
+"""Expected results for estimators and utilities to generate them."""
