+++ aeon/testing/tests/test_all_estimators.py
@@ -24,7 +24,7 @@ from aeon.exceptions import NotFittedError
 from aeon.forecasting.base import BaseForecaster
 from aeon.registry import all_estimators
 from aeon.regression.deep_learning.base import BaseDeepRegressor
-from aeon.tests.test_config import (
+from aeon.testing.tests.test_config import (
     EXCLUDE_ESTIMATORS,
     EXCLUDED_TESTS,
     NON_STATE_CHANGING_METHODS,
