+++ aeon/transformations/hierarchical/tests/test_reconcile.py
@@ -8,7 +8,7 @@ from pandas.testing import assert_frame_equal
 
 from aeon.forecasting.base import ForecastingHorizon
 from aeon.forecasting.exp_smoothing import ExponentialSmoothing
-from aeon.tests.test_all_estimators import PR_TESTING
+from aeon.testing import PR_TESTING
 from aeon.transformations.hierarchical.aggregate import Aggregator
 from aeon.transformations.hierarchical.reconcile import Reconciler
 from aeon.utils._testing.hierarchical import _bottom_hier_datagen
